backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
load-dotenv==0.1.0
mysqlclient==2.2.6
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
